Replaced float in favor for decimal in models (better precision)

+++ Fortnox.NET/Models/Offer/Offer.cs
@@ -130,7 +130,7 @@ namespace FortnoxNET.Models.Offer
         public bool Sent { get; set; }
 
         [JsonReadOnly]
-        public float? TaxReduction { get; set; }
+        public decimal? TaxReduction { get; set; }
         
         public string TermsOfDelivery { get; set; }
